Fixed camera jitter
This is not an inclusive fix, it helps the issue (by using doubles) but
at extreme distances everything is going to break again. We need to
"Treadmill" the world to keep everything relatively close to the origin

+++ app/Entities/Entity.h
@@ -29,7 +29,7 @@ public:
 	//Rotation Logic
 	virtual void setRotation(osg::Vec3d);
 	void modRotation(osg::Vec3d);
-	osg::Vec3f getRotation() { return _rotation; }
+	osg::Vec3d getRotation() { return _rotation; }
 
 	//OSG Node Logic
 	virtual void attach(osg::Group* group);
